make lightning.qubit2 the new lightning.qubit
